backend/event-service/pyproject.toml
[tool.poetry]
name = "event-service"
version = "0.1.0"
description = ""
authors = ["ionic101 <iv.osipov2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
dotenv = "^0.9.9"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
